wimpt_messages preventing program info appearing from the menu
The program info dialogue box was not appearing when hovering the mouse over
the "Info" entry in the menu.  I suspect that a WIMP message is being blocked
by wimpt_messages and will investigate this later.  I'm pushing a quick fix
(commenting out the call to wimpt_messages) mostly to try out pushing an
update to github.